LayoutTests/media/modern-media-controls/macos-inline-media-controls/mac-inline-controls-time-control-no-overlap-expected.txt

Testing that time control doesn't overlap right button container in inline media controls.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS timeControlRect.right <= rightContainerRect.left is true
PASS successfullyParsed is true

TEST COMPLETE
